[Yield Farm] Tidy up typings and error checks
